- update NFTs on wallet page creation
